content/media/test/test_load_candidates.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=465458
-->
<head>
  <title>Test for Bug 465458</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=465458">Mozilla Bug 465458</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 465458 **/

var videos = [];

function finish(evt) {
  is(evt.target._error, false, "Shouldn't have thrown an error - " + evt.target._name);
  evt.target._finished = true;
  if (AllFinished(videos)) {
    SimpleTest.finish();
  }
}

function errorHandler(evt) {
  evt.stopPropagation();
  evt.target._error = true;
  finish(evt);
}

var extenstion = {
  "audio/wav" : "wav",
  "audio/x-wav": "wav",
  "video/ogg" : "ogv",
  "audio/ogg" : "oga",
  "video/webm" : "webm"
};

for (var i=0; i<gSmallTests.length; ++i) {
  var v = document.createElement('video');
  var test = gSmallTests[i];
  if (!v.canPlayType(test.type))
    continue;
  v._error = false;
  v._finished = false;
  v._name = test.name;
  
  var s1 = document.createElement("source");
  s1.type = test.type;
  s1.src = "404." + extenstion[test.type];
  v.appendChild(s1);
  
  var s2 = document.createElement("source");
  s2.type = test.type;
  s2.src = test.name;
  v.appendChild(s2);
  
  v.addEventListener("error", errorHandler, false);
  v.addEventListener("loadeddata", finish, false);
  
  document.body.appendChild(v);
  
  videos.push(v);
}


if (videos.length == 0) {
  todo(false, "No types supported");
} else {
  SimpleTest.waitForExplicitFinish();
}

</script>
</pre>
</body>
</html>
